Decompose MeshScene into focused viewport rendering layers
MeshScene.tsx had grown to 1242 lines mixing geometry tessellation
display, FEM mesh rendering, result colormapping, and BC/load
visualisation in one component. Split it into layers, each receiving
the shared topology plus only the store slices it needs:

- GeometryLayer: OCCT tessellation of the STEP shape (display only)
- FemMeshLayer: boundary surface + edge wireframes, deformed variants
- ResultsColormap: result-field vertex colours on the deformed surface
- BoundaryConditionLayer: BC/load face highlights, markers, load arrows
- useMeshTopology: shared node map, model extent, boundary face
  extraction and pickable boundary-mesh topology, computed once
- useFacePick: face-pick click handler shared by the pickable surfaces

MeshScene is now a ~90-line compositor that mounts layers based on the
active mode and representation. Rendering behaviour is unchanged.

Closes #171
