+++ web/src/components/viewport/MeshScene.tsx
@@ -1,190 +1,36 @@
 // SPDX-FileCopyrightText: 2026 Michael Kofler
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
+// Thin compositor for the viewport scene. Derives the shared mesh topology and
+// deformation scale once, then mounts the rendering layers based on the active
+// mode / representation:
+//   GeometryLayer          — OCCT tessellation of the STEP shape (display only)
+//   FemMeshLayer           — FEM boundary surface + edge wireframes (± deformed)
+//   ResultsColormap        — result-field vertex colours on the deformed surface
+//   BoundaryConditionLayer — BC/load face highlights, markers and load arrows
+
 import { useMemo } from "react";
-import * as THREE from "three";
-import type { ThreeEvent } from "@react-three/fiber";
-import {
-  useModelStore,
-  loadKind,
-  loadComponents,
-} from "../../store/modelStore";
-import type { Node, ResultType } from "../../store/modelStore";
-import {
-  buildBoundaryMeshTopo,
-  pickFaceNodeIds,
-  toggleFaceSelection,
-} from "../../lib/facePick";
-import type { Vec3, Tri } from "../../lib/facePick";
-import { nodeVonMisesField } from "../../lib/resultField";
+import { useModelStore } from "../../store/modelStore";
+import { useMeshTopology } from "./useMeshTopology";
+import { useFacePick } from "./useFacePick";
+import { GeometryLayer } from "./GeometryLayer";
+import { FemMeshLayer } from "./FemMeshLayer";
+import { ResultsColormap } from "./ResultsColormap";
+import { BoundaryConditionLayer } from "./BoundaryConditionLayer";
 
 const TARGET_DEFORM_FRACTION = 0.2;
 
-// ── CHEXA geometry ────────────────────────────────────────────────────────────
-
-const HEX_EDGES: [number, number][] = [
-  [0, 1],
-  [1, 2],
-  [2, 3],
-  [3, 0],
-  [4, 5],
-  [5, 6],
-  [6, 7],
-  [7, 4],
-  [0, 4],
-  [1, 5],
-  [2, 6],
-  [3, 7],
-];
-
-const HEX_FACE_DEFS: [number, number, number, number][] = [
-  [0, 1, 2, 3],
-  [4, 5, 6, 7],
-  [0, 1, 5, 4],
-  [2, 3, 7, 6],
-  [0, 3, 7, 4],
-  [1, 2, 6, 5],
-];
-
-function extractBoundaryQuadFaceIds(
-  hexElems: { nodeIds: number[] }[],
-): [number, number, number, number][] {
-  const faceMap = new Map<
-    string,
-    { face: [number, number, number, number]; count: number }
-  >();
-  for (const el of hexElems) {
-    for (const [a, b, c, d] of HEX_FACE_DEFS) {
-      const face: [number, number, number, number] = [
-        el.nodeIds[a],
-        el.nodeIds[b],
-        el.nodeIds[c],
-        el.nodeIds[d],
-      ];
-      const key = [...face].sort((x, y) => x - y).join(",");
-      const entry = faceMap.get(key);
-      if (entry) {
-        entry.count++;
-      } else {
-        faceMap.set(key, { face, count: 1 });
-      }
-    }
-  }
-  return [...faceMap.values()].filter((e) => e.count === 1).map((e) => e.face);
-}
-
-// ── CTETRA geometry ───────────────────────────────────────────────────────────
-
-const TET_EDGES: [number, number][] = [
-  [0, 1],
-  [0, 2],
-  [0, 3],
-  [1, 2],
-  [1, 3],
-  [2, 3],
-];
-
-const TET_FACE_DEFS: [number, number, number][] = [
-  [0, 1, 2],
-  [0, 1, 3],
-  [0, 2, 3],
-  [1, 2, 3],
-];
-
-function extractBoundaryTriFaceIds(
-  tetElems: { nodeIds: number[] }[],
-): [number, number, number][] {
-  const faceMap = new Map<
-    string,
-    { face: [number, number, number]; count: number }
-  >();
-  for (const el of tetElems) {
-    for (const [a, b, c] of TET_FACE_DEFS) {
-      const face: [number, number, number] = [
-        el.nodeIds[a],
-        el.nodeIds[b],
-        el.nodeIds[c],
-      ];
-      const key = [...face].sort((x, y) => x - y).join(",");
-      const entry = faceMap.get(key);
-      if (entry) {
-        entry.count++;
-      } else {
-        faceMap.set(key, { face, count: 1 });
-      }
-    }
-  }
-  return [...faceMap.values()].filter((e) => e.count === 1).map((e) => e.face);
-}
-
-// ── Helpers ───────────────────────────────────────────────────────────────────
-
-function buildFacePositions(
-  nodeIds: number[],
-  triangles: [number, number, number][],
-  nodeMap: Map<number, { n: Node; i: number }>,
-): Float32Array | null {
-  const nodeIdSet = new Set(nodeIds);
-  const positions: number[] = [];
-  for (const [a, b, c] of triangles) {
-    if (!nodeIdSet.has(a) || !nodeIdSet.has(b) || !nodeIdSet.has(c)) continue;
-    const na = nodeMap.get(a)?.n,
-      nb = nodeMap.get(b)?.n,
-      nc = nodeMap.get(c)?.n;
-    if (!na || !nb || !nc) continue;
-    positions.push(na.x, na.y, na.z, nb.x, nb.y, nb.z, nc.x, nc.y, nc.z);
-  }
-  return positions.length > 0 ? new Float32Array(positions) : null;
-}
-
-// ── Component ─────────────────────────────────────────────────────────────────
-
 export function MeshScene() {
   const nodes = useModelStore((s) => s.nodes);
-  const elements = useModelStore((s) => s.elements);
-  const constraints = useModelStore((s) => s.constraints);
   const result = useModelStore((s) => s.result);
-  const resultType = useModelStore((s) => s.resultType);
   const mode = useModelStore((s) => s.mode);
   const stepSurface = useModelStore((s) => s.stepSurface);
   const deformScaleFactor = useModelStore((s) => s.deformScale);
-  const surfaceTriangles = useModelStore((s) => s.surfaceTriangles);
-  const surfaceFaceIds = useModelStore((s) => s.surfaceFaceIds);
   const viewRepr = useModelStore((s) => s.viewRepr);
-  const pickMode = useModelStore((s) => s.pickMode);
-  const pickTargetGroupId = useModelStore((s) => s.pickTargetGroupId);
-  const selectedFace = useModelStore((s) => s.selectedFace);
-  const pendingFaces = useModelStore((s) => s.pendingFaces);
-  const setSelectedFace = useModelStore((s) => s.setSelectedFace);
-  const setPendingFaces = useModelStore((s) => s.setPendingFaces);
-  const bcGroups = useModelStore((s) => s.bcGroups);
-  const loadGroups = useModelStore((s) => s.loadGroups);
-  const loadDisplay = useModelStore((s) => s.loadDisplay);
-  const showUndeformedOverlay = useModelStore((s) => s.showUndeformedOverlay);
 
-  const nodeMap = useMemo(
-    () => new Map(nodes.map((n, i) => [n.id, { n, i }])),
-    [nodes],
-  );
-
-  const modelSize = useMemo(() => {
-    if (nodes.length === 0) return 1;
-    let minX = Infinity,
-      maxX = -Infinity,
-      minY = Infinity,
-      maxY = -Infinity,
-      minZ = Infinity,
-      maxZ = -Infinity;
-    for (const n of nodes) {
-      if (n.x < minX) minX = n.x;
-      if (n.x > maxX) maxX = n.x;
-      if (n.y < minY) minY = n.y;
-      if (n.y > maxY) maxY = n.y;
-      if (n.z < minZ) minZ = n.z;
-      if (n.z > maxZ) maxZ = n.z;
-    }
-    return Math.max(maxX - minX, maxY - minY, maxZ - minZ, 1e-9);
-  }, [nodes]);
+  const topology = useMeshTopology();
+  const onFacePick = useFacePick(topology.boundaryMeshTopo);
+  const { modelSize } = topology;
 
   // Automatic fit-to-view scale (max displacement → TARGET_DEFORM_FRACTION of the
   // model size) multiplied by the user-controlled deformScaleFactor. A factor of
@@ -200,750 +46,8 @@ export function MeshScene() {
     return ((TARGET_DEFORM_FRACTION * modelSize) / maxDisp) * deformScaleFactor;
   }, [result, modelSize, deformScaleFactor]);
 
-  const hexElements = useMemo(
-    () => elements.filter((e) => e.type === "CHEXA"),
-    [elements],
-  );
-  const tetElements = useMemo(
-    () => elements.filter((e) => e.type === "CTETRA"),
-    [elements],
-  );
-
-  const boundaryQuadFaceIds = useMemo(
-    () => extractBoundaryQuadFaceIds(hexElements),
-    [hexElements],
-  );
-  const boundaryTriFaceIds = useMemo(
-    () => extractBoundaryTriFaceIds(tetElements),
-    [tetElements],
-  );
-
-  // Boundary mesh topology for face picking.
-  // Triangle order matches the undeformedSurface BufferGeometry exactly so that
-  // e.faceIndex from raycasting maps directly into this triangles array.
-  //
-  // When surfaceFaceIds from the store is present (STEP mesh via Netgen OCC),
-  // a sorted-vertex lookup maps each boundary triangle to its OCC face index.
-  // pickFaceNodeIds then does an instant lookup instead of BFS flood-fill.
-  const boundaryMeshTopo = useMemo(() => {
-    const triangles: Tri[] = [];
-    for (const [a, b, c, d] of boundaryQuadFaceIds) {
-      triangles.push([a, b, c], [a, c, d]);
-    }
-    for (const [a, b, c] of boundaryTriFaceIds) {
-      triangles.push([a, b, c]);
-    }
-    if (triangles.length === 0) return null;
-
-    const getPos = (id: number): Vec3 => {
-      const n = nodeMap.get(id)?.n;
-      return n ? [n.x, n.y, n.z] : [0, 0, 0];
-    };
-
-    // Build per-boundary-triangle face IDs by matching sorted vertex triples.
-    // surfaceTriangles / surfaceFaceIds come from Netgen in surface-element order
-    // which differs from the tet boundary extraction order, so a direct index
-    // mapping would be wrong.  Sorting the three vertex IDs of each triangle
-    // produces an order-independent key that matches across both orderings.
-    let faceIds: number[] | undefined;
-    if (
-      surfaceTriangles &&
-      surfaceFaceIds &&
-      surfaceTriangles.length === surfaceFaceIds.length
-    ) {
-      const sk = (p: number, q: number, r: number): string =>
-        [p, q, r].sort((x, y) => x - y).join(",");
-      const keyToFaceId = new Map<string, number>();
-      for (let i = 0; i < surfaceTriangles.length; i++) {
-        const [a, b, c] = surfaceTriangles[i];
-        keyToFaceId.set(sk(a, b, c), surfaceFaceIds[i]);
-      }
-      const mapped = triangles.map(([a, b, c]) => keyToFaceId.get(sk(a, b, c)));
-      if (mapped.every((id) => id !== undefined)) {
-        faceIds = mapped as number[];
-      }
-    }
-
-    return buildBoundaryMeshTopo(triangles, getPos, faceIds);
-  }, [
-    boundaryQuadFaceIds,
-    boundaryTriFaceIds,
-    nodeMap,
-    surfaceTriangles,
-    surfaceFaceIds,
-  ]);
-
-  const undeformedEdgePositions = useMemo(() => {
-    const segs: number[] = [];
-    const coord = (id: number) => {
-      const e = nodeMap.get(id)!;
-      return [e.n.x, e.n.y, e.n.z];
-    };
-    for (const el of hexElements) {
-      for (const [a, b] of HEX_EDGES) {
-        segs.push(...coord(el.nodeIds[a]), ...coord(el.nodeIds[b]));
-      }
-    }
-    for (const el of tetElements) {
-      for (const [a, b] of TET_EDGES) {
-        segs.push(...coord(el.nodeIds[a]), ...coord(el.nodeIds[b]));
-      }
-    }
-    return segs.length > 0 ? new Float32Array(segs) : null;
-  }, [hexElements, tetElements, nodeMap]);
-
-  const undeformedSurfaceEdgePositions = useMemo(() => {
-    const seen = new Set<string>();
-    const segs: number[] = [];
-    const addEdge = (a: number, b: number) => {
-      const key = a < b ? `${a},${b}` : `${b},${a}`;
-      if (seen.has(key)) return;
-      seen.add(key);
-      const na = nodeMap.get(a)?.n,
-        nb = nodeMap.get(b)?.n;
-      if (!na || !nb) return;
-      segs.push(na.x, na.y, na.z, nb.x, nb.y, nb.z);
-    };
-    for (const [a, b, c, d] of boundaryQuadFaceIds) {
-      addEdge(a, b);
-      addEdge(b, c);
-      addEdge(c, d);
-      addEdge(d, a);
-    }
-    for (const [a, b, c] of boundaryTriFaceIds) {
-      addEdge(a, b);
-      addEdge(b, c);
-      addEdge(c, a);
-    }
-    return segs.length > 0 ? new Float32Array(segs) : null;
-  }, [boundaryQuadFaceIds, boundaryTriFaceIds, nodeMap]);
-
-  const deformedEdgePositions = useMemo(() => {
-    if (!result) return null;
-    const d = result.displacements;
-    const coord = (id: number) => {
-      const { n, i } = nodeMap.get(id)!;
-      return [
-        n.x + (d[i * 3] ?? 0) * deformScale,
-        n.y + (d[i * 3 + 1] ?? 0) * deformScale,
-        n.z + (d[i * 3 + 2] ?? 0) * deformScale,
-      ];
-    };
-    const segs: number[] = [];
-    for (const el of hexElements) {
-      for (const [a, b] of HEX_EDGES) {
-        segs.push(...coord(el.nodeIds[a]), ...coord(el.nodeIds[b]));
-      }
-    }
-    for (const el of tetElements) {
-      for (const [a, b] of TET_EDGES) {
-        segs.push(...coord(el.nodeIds[a]), ...coord(el.nodeIds[b]));
-      }
-    }
-    return segs.length > 0 ? new Float32Array(segs) : null;
-  }, [result, hexElements, tetElements, nodeMap, deformScale]);
-
-  // Boundary (surface) edges of the deformed mesh — the Surface representation in
-  // results mode. Mirrors undeformedSurfaceEdgePositions but on deformed coords.
-  const deformedSurfaceEdgePositions = useMemo(() => {
-    if (!result) return null;
-    const d = result.displacements;
-    const seen = new Set<string>();
-    const segs: number[] = [];
-    const addEdge = (a: number, b: number) => {
-      const key = a < b ? `${a},${b}` : `${b},${a}`;
-      if (seen.has(key)) return;
-      seen.add(key);
-      const na = nodeMap.get(a),
-        nb = nodeMap.get(b);
-      if (!na || !nb) return;
-      segs.push(
-        na.n.x + (d[na.i * 3] ?? 0) * deformScale,
-        na.n.y + (d[na.i * 3 + 1] ?? 0) * deformScale,
-        na.n.z + (d[na.i * 3 + 2] ?? 0) * deformScale,
-        nb.n.x + (d[nb.i * 3] ?? 0) * deformScale,
-        nb.n.y + (d[nb.i * 3 + 1] ?? 0) * deformScale,
-        nb.n.z + (d[nb.i * 3 + 2] ?? 0) * deformScale,
-      );
-    };
-    for (const [a, b, c, dd] of boundaryQuadFaceIds) {
-      addEdge(a, b);
-      addEdge(b, c);
-      addEdge(c, dd);
-      addEdge(dd, a);
-    }
-    for (const [a, b, c] of boundaryTriFaceIds) {
-      addEdge(a, b);
-      addEdge(b, c);
-      addEdge(c, a);
-    }
-    return segs.length > 0 ? new Float32Array(segs) : null;
-  }, [result, boundaryQuadFaceIds, boundaryTriFaceIds, nodeMap, deformScale]);
-
-  // Per-node von Mises: volume-weighted average of the element-level stresses,
-  // shared with the colorbar range in resultField so the two stay identical.
-  const nodeVonMises = useMemo(
-    () => (result ? nodeVonMisesField(result, nodes, elements) : null),
-    [result, nodes, elements],
-  );
-
-  const deformedSurface = useMemo(() => {
-    if (!result) return null;
-    const hasQuads = boundaryQuadFaceIds.length > 0;
-    const hasTris = boundaryTriFaceIds.length > 0;
-    if (!hasQuads && !hasTris) return null;
-
-    const d = result.displacements;
-
-    // Compute per-node scalar value for the selected result type
-    const nodeValue = (i: number, rt: ResultType): number => {
-      switch (rt) {
-        case "Ux":
-          return d[i * 3] ?? 0;
-        case "Uy":
-          return d[i * 3 + 1] ?? 0;
-        case "Uz":
-          return d[i * 3 + 2] ?? 0;
-        case "Von Mises stress":
-          return nodeVonMises?.[i] ?? 0;
-        default: {
-          const ux = d[i * 3] ?? 0,
-            uy = d[i * 3 + 1] ?? 0,
-            uz = d[i * 3 + 2] ?? 0;
-          return Math.sqrt(ux * ux + uy * uy + uz * uz);
-        }
-      }
-    };
-
-    let minVal = Infinity,
-      maxVal = -Infinity;
-    for (let i = 0; i < nodes.length; i++) {
-      const v = nodeValue(i, resultType);
-      if (v < minVal) minVal = v;
-      if (v > maxVal) maxVal = v;
-    }
-    const range = maxVal - minVal || 1;
-
-    const positions: number[] = [];
-    const colors: number[] = [];
-
-    const deformedPos = (id: number): [number, number, number] => {
-      const { n, i } = nodeMap.get(id)!;
-      return [
-        n.x + (d[i * 3] ?? 0) * deformScale,
-        n.y + (d[i * 3 + 1] ?? 0) * deformScale,
-        n.z + (d[i * 3 + 2] ?? 0) * deformScale,
-      ];
-    };
-    const nodeColor = (id: number): [number, number, number] => {
-      const { i } = nodeMap.get(id)!;
-      const t = (nodeValue(i, resultType) - minVal) / range;
-      const c = new THREE.Color();
-      c.setHSL(0.667 * (1 - t), 1, 0.5);
-      return [c.r, c.g, c.b];
-    };
-
-    for (const [a, b, c_, dd] of boundaryQuadFaceIds) {
-      const pa = deformedPos(a),
-        pb = deformedPos(b),
-        pc = deformedPos(c_),
-        pd = deformedPos(dd);
-      const ca = nodeColor(a),
-        cb = nodeColor(b),
-        cc = nodeColor(c_),
-        cd = nodeColor(dd);
-      positions.push(...pa, ...pb, ...pc, ...pa, ...pc, ...pd);
-      colors.push(...ca, ...cb, ...cc, ...ca, ...cc, ...cd);
-    }
-    for (const [a, b, c_] of boundaryTriFaceIds) {
-      const pa = deformedPos(a),
-        pb = deformedPos(b),
-        pc = deformedPos(c_);
-      const ca = nodeColor(a),
-        cb = nodeColor(b),
-        cc = nodeColor(c_);
-      positions.push(...pa, ...pb, ...pc);
-      colors.push(...ca, ...cb, ...cc);
-    }
-
-    return {
-      positions: new Float32Array(positions),
-      colors: new Float32Array(colors),
-    };
-  }, [
-    result,
-    resultType,
-    nodeVonMises,
-    boundaryQuadFaceIds,
-    boundaryTriFaceIds,
-    nodeMap,
-    nodes,
-    deformScale,
-  ]);
-
-  // Undeformed surface for face picking
-  const undeformedSurface = useMemo(() => {
-    const hasQuads = boundaryQuadFaceIds.length > 0;
-    const hasTris = boundaryTriFaceIds.length > 0;
-    if (!hasQuads && !hasTris) return null;
-
-    const positions: number[] = [];
-    const normals: number[] = [];
-
-    const p = (n: { x: number; y: number; z: number }) =>
-      [n.x, n.y, n.z] as [number, number, number];
-
-    for (const [a, b, c_, dd] of boundaryQuadFaceIds) {
-      const pa = nodeMap.get(a)!.n,
-        pb = nodeMap.get(b)!.n;
-      const pc = nodeMap.get(c_)!.n,
-        pd = nodeMap.get(dd)!.n;
-      positions.push(
-        ...p(pa),
-        ...p(pb),
-        ...p(pc),
-        ...p(pa),
-        ...p(pc),
-        ...p(pd),
-      );
-      const AB = new THREE.Vector3(pb.x - pa.x, pb.y - pa.y, pb.z - pa.z);
-      const AC = new THREE.Vector3(pc.x - pa.x, pc.y - pa.y, pc.z - pa.z);
-      const norm = AB.cross(AC).normalize();
-      for (let k = 0; k < 6; k++) normals.push(norm.x, norm.y, norm.z);
-    }
-    for (const [a, b, c_] of boundaryTriFaceIds) {
-      const pa = nodeMap.get(a)!.n,
-        pb = nodeMap.get(b)!.n,
-        pc = nodeMap.get(c_)!.n;
-      positions.push(...p(pa), ...p(pb), ...p(pc));
-      const AB = new THREE.Vector3(pb.x - pa.x, pb.y - pa.y, pb.z - pa.z);
-      const AC = new THREE.Vector3(pc.x - pa.x, pc.y - pa.y, pc.z - pa.z);
-      const norm = AB.cross(AC).normalize();
-      for (let k = 0; k < 3; k++) normals.push(norm.x, norm.y, norm.z);
-    }
-
-    return {
-      positions: new Float32Array(positions),
-      normals: new Float32Array(normals),
-    };
-  }, [boundaryQuadFaceIds, boundaryTriFaceIds, nodeMap]);
-
-  // Selected face highlight — the current (latest) picked face in the active pick session.
-  const selectedFacePositions = useMemo(() => {
-    if (!selectedFace || !boundaryMeshTopo) return null;
-    return buildFacePositions(
-      selectedFace.nodeIds,
-      boundaryMeshTopo.triangles,
-      nodeMap,
-    );
-  }, [selectedFace, boundaryMeshTopo, nodeMap]);
-
-  // Pending faces — accumulated via shift-click during this pick session.
-  const pendingFacePositions = useMemo(() => {
-    if (pendingFaces.length === 0 || !boundaryMeshTopo) return null;
-    const allNodeIds = pendingFaces.flatMap((f) => f.nodeIds);
-    return buildFacePositions(allNodeIds, boundaryMeshTopo.triangles, nodeMap);
-  }, [pendingFaces, boundaryMeshTopo, nodeMap]);
-
-  // BC face highlights — all committed faces across all BC groups.
-  // When in pick mode targeting a specific group, that group is highlighted separately (below).
-  const bcFaceHighlights = useMemo(() => {
-    if (!boundaryMeshTopo) return null;
-    return bcGroups
-      .flatMap((g) =>
-        g.faces.map((f) => ({
-          groupId: g.id,
-          positions: buildFacePositions(
-            f.nodeIds,
-            boundaryMeshTopo.triangles,
-            nodeMap,
-          ),
-        })),
-      )
-      .filter((h) => h.positions !== null) as {
-      groupId: number;
-      positions: Float32Array;
-    }[];
-  }, [bcGroups, boundaryMeshTopo, nodeMap]);
-
-  // Load face highlights — all committed faces across all load groups.
-  const loadFaceHighlights = useMemo(() => {
-    if (!boundaryMeshTopo) return null;
-    return loadGroups
-      .flatMap((g) =>
-        g.faces.map((f) => ({
-          groupId: g.id,
-          positions: buildFacePositions(
-            f.nodeIds,
-            boundaryMeshTopo.triangles,
-            nodeMap,
-          ),
-        })),
-      )
-      .filter((h) => h.positions !== null) as {
-      groupId: number;
-      positions: Float32Array;
-    }[];
-  }, [loadGroups, boundaryMeshTopo, nodeMap]);
-
-  // Face picking handler.
-  // When OCC face IDs are available (STEP mesh via Netgen), uses instant face ID
-  // lookup — topologically exact, works on any curved or flat CAD face.
-  // Falls back to BFS flood-fill with normal-angle thresholds when no face IDs
-  // are present (parametric box mesh or .inp import).
-  function handleFacePick(e: ThreeEvent<MouseEvent>) {
-    if (!pickMode || e.faceIndex == null || !boundaryMeshTopo) return;
-    e.stopPropagation();
-
-    const startIdx = e.faceIndex;
-    if (startIdx >= boundaryMeshTopo.triangles.length) return;
-
-    const faceNodeIds = [...pickFaceNodeIds(startIdx, boundaryMeshTopo)];
-    const normal =
-      e.face?.normal.clone().normalize() ?? new THREE.Vector3(0, 1, 0);
-    const ax = Math.abs(normal.x),
-      ay = Math.abs(normal.y),
-      az = Math.abs(normal.z);
-    let axis: "X" | "Y" | "Z";
-    let isMax: boolean;
-    if (ax >= ay && ax >= az) {
-      axis = "X";
-      isMax = normal.x > 0;
-    } else if (ay >= ax && ay >= az) {
-      axis = "Y";
-      isMax = normal.y > 0;
-    } else {
-      axis = "Z";
-      isMax = normal.z > 0;
-    }
-
-    // Current selection = accumulated pending faces plus the active one.
-    // Re-selecting an already-picked face toggles it off instead of adding a
-    // duplicate (#264).
-    const current = selectedFace
-      ? [...pendingFaces, selectedFace]
-      : pendingFaces;
-    const next = toggleFaceSelection(current, {
-      nodeIds: faceNodeIds,
-      axis,
-      isMax,
-      label: "",
-    });
-
-    // Re-split into pending faces + the active (last) selection.
-    setPendingFaces(next.slice(0, -1));
-    setSelectedFace(next.length > 0 ? next[next.length - 1] : null);
-  }
-
-  // BC marker: centroid + quaternion that orients triangles outward from the model
-  const bcMarkerData = useMemo(() => {
-    if (constraints.length === 0 || nodes.length === 0) return null;
-    const ids = new Set(constraints.map((c) => c.nodeId));
-    let cx = 0,
-      cy = 0,
-      cz = 0,
-      count = 0;
-    for (const id of ids) {
-      const e = nodeMap.get(id);
-      if (e) {
-        cx += e.n.x;
-        cy += e.n.y;
-        cz += e.n.z;
-        count++;
-      }
-    }
-    if (count === 0) return null;
-    cx /= count;
-    cy /= count;
-    cz /= count;
-
-    // Outward normal: direction from model centroid toward BC face centroid
-    let mx = 0,
-      my = 0,
-      mz = 0;
-    for (const n of nodes) {
-      mx += n.x;
-      my += n.y;
-      mz += n.z;
-    }
-    mx /= nodes.length;
-    my /= nodes.length;
-    mz /= nodes.length;
-    const dx = cx - mx,
-      dy = cy - my,
-      dz = cz - mz;
-    const len = Math.sqrt(dx * dx + dy * dy + dz * dz) || 1;
-    const outward = new THREE.Vector3(dx / len, dy / len, dz / len);
-
-    // Rotate group so that -Y (cone base direction) aligns with outward normal
-    const q = new THREE.Quaternion();
-    q.setFromUnitVectors(new THREE.Vector3(0, -1, 0), outward);
-    return { pos: [cx, cy, cz] as [number, number, number], quaternion: q };
-  }, [constraints, nodeMap, nodes]);
-
-  // Load glyphs — one arrow per force/pressure group, at the centroid of its
-  // loaded nodes. Force arrows point along the applied force; pressure arrows
-  // point into the loaded face (positive pressure pushes inward). Driven by
-  // loadGroups because force/pressure loads reach the solver as surface tractions
-  // rather than nodal forces. Moments carry no single resultant, so they are
-  // skipped.
-  const loadArrows = useMemo(() => {
-    if (loadGroups.length === 0 || nodes.length === 0) return [];
-    // Model centroid — used to orient pressure arrows inward.
-    let mx = 0,
-      my = 0,
-      mz = 0;
-    for (const n of nodes) {
-      mx += n.x;
-      my += n.y;
-      mz += n.z;
-    }
-    mx /= nodes.length;
-    my /= nodes.length;
-    mz /= nodes.length;
-
-    const up = new THREE.Vector3(0, 1, 0);
-    const arrows: {
-      pos: [number, number, number];
-      quaternion: THREE.Quaternion;
-    }[] = [];
-    for (const g of loadGroups) {
-      const kind = loadKind(g);
-      if (kind === "moment") continue;
-      let cx = 0,
-        cy = 0,
-        cz = 0,
-        count = 0;
-      for (const f of g.faces) {
-        for (const id of f.nodeIds) {
-          const e = nodeMap.get(id);
-          if (e) {
-            cx += e.n.x;
-            cy += e.n.y;
-            cz += e.n.z;
-            count++;
-          }
-        }
-      }
-      if (count === 0) continue;
-      cx /= count;
-      cy /= count;
-      cz /= count;
-
-      let dir: THREE.Vector3;
-      if (kind === "pressure") {
-        dir = new THREE.Vector3(mx - cx, my - cy, mz - cz);
-      } else {
-        const vec = loadComponents(g);
-        dir = new THREE.Vector3(vec[0], vec[1], vec[2]);
-      }
-      if (dir.lengthSq() < 1e-30) continue;
-      dir.normalize();
-      const q = new THREE.Quaternion();
-      q.setFromUnitVectors(up, dir);
-      arrows.push({ pos: [cx, cy, cz], quaternion: q });
-    }
-    return arrows;
-  }, [loadGroups, nodes, nodeMap]);
-
-  // Per-node load glyphs — one arrow at every loaded node, sized by the
-  // work-equivalent load that node carries: its tributary-area share of the
-  // group total (force ⇒ totalForce·Aᵢ/A, pressure ⇒ p·Aᵢ, both in N). This is
-  // what actually reaches the solver as a surface traction, in contrast to the
-  // single statically-equivalent resultant. Shown when loadDisplay === "nodal".
-  // Force arrows point along the applied force; pressure arrows point into the
-  // surface along the per-node inward normal. Moments carry no resultant force
-  // and are skipped, matching the resultant view (issue #196).
-  const nodalLoadArrows = useMemo(() => {
-    if (loadGroups.length === 0 || nodes.length === 0) return [];
-
-    // Model centroid — used to orient pressure arrows inward.
-    let mx = 0,
-      my = 0,
-      mz = 0;
-    for (const n of nodes) {
-      mx += n.x;
-      my += n.y;
-      mz += n.z;
-    }
-    mx /= nodes.length;
-    my /= nodes.length;
-    mz /= nodes.length;
-    const modelCentroid = new THREE.Vector3(mx, my, mz);
-
-    const posOf = (id: number): THREE.Vector3 | null => {
-      const e = nodeMap.get(id);
-      return e ? new THREE.Vector3(e.n.x, e.n.y, e.n.z) : null;
-    };
-
-    // Accumulated per node across all groups: load direction (unit) and
-    // magnitude (N). A node shared by two groups gets the larger arrow.
-    type NodalLoad = { dir: THREE.Vector3; mag: number };
-    const byNode = new Map<number, NodalLoad>();
-
-    for (const g of loadGroups) {
-      const kind = loadKind(g);
-      if (kind === "moment") continue;
-
-      const nodeSet = new Set<number>();
-      for (const f of g.faces) for (const id of f.nodeIds) nodeSet.add(id);
-      if (nodeSet.size === 0) continue;
-
-      // Tributary area per node and (for pressure) accumulated inward normal,
-      // integrated over the boundary faces that lie entirely on the group's
-      // selected faces — the same membership test the solver uses to apply the
-      // surface traction, so the per-node share matches the FE load.
-      const tribArea = new Map<number, number>();
-      const inward = new Map<number, THREE.Vector3>();
-      let totalArea = 0;
-
-      const addFace = (ids: number[]) => {
-        if (!ids.every((id) => nodeSet.has(id))) return;
-        const pts = ids.map(posOf);
-        if (pts.some((p) => p === null)) return;
-        const p = pts as THREE.Vector3[];
-
-        // Area + outward normal: triangle directly, quad via its diagonals.
-        let area: number;
-        const normal = new THREE.Vector3();
-        if (p.length === 3) {
-          normal
-            .copy(p[1])
-            .sub(p[0])
-            .cross(new THREE.Vector3().copy(p[2]).sub(p[0]));
-          area = 0.5 * normal.length();
-        } else {
-          normal
-            .copy(p[2])
-            .sub(p[0])
-            .cross(new THREE.Vector3().copy(p[3]).sub(p[1]));
-          area = 0.5 * normal.length();
-        }
-        if (area < 1e-30) return;
-        normal.normalize();
-        totalArea += area;
-
-        // Orient the face normal inward (positive pressure pushes in).
-        const fc = new THREE.Vector3();
-        for (const v of p) fc.add(v);
-        fc.divideScalar(p.length);
-        if (normal.dot(new THREE.Vector3().copy(modelCentroid).sub(fc)) < 0)
-          normal.negate();
-
-        const share = area / p.length;
-        for (const id of ids) {
-          tribArea.set(id, (tribArea.get(id) ?? 0) + share);
-          if (kind === "pressure") {
-            const acc = inward.get(id) ?? new THREE.Vector3();
-            inward.set(id, acc.addScaledVector(normal, area));
-          }
-        }
-      };
-
-      for (const tri of boundaryTriFaceIds) addFace(tri);
-      for (const quad of boundaryQuadFaceIds) addFace(quad);
-      if (totalArea < 1e-30) continue;
-
-      // Force direction is shared by every node; pressure direction is per-node.
-      const force = loadComponents(g);
-      const forceMag = Math.hypot(force[0], force[1], force[2]);
-      const forceDir = new THREE.Vector3();
-      if (kind === "force") {
-        if (forceMag < 1e-30) continue;
-        forceDir.set(force[0], force[1], force[2]).normalize();
-      }
-      // Pressure stores its (signed) magnitude in totalForce; positive pushes in.
-      const pressureSign = g.totalForce < 0 ? -1 : 1;
-
-      for (const [id, a] of tribArea) {
-        let dir: THREE.Vector3;
-        let mag: number;
-        if (kind === "force") {
-          dir = forceDir;
-          mag = forceMag * (a / totalArea);
-        } else {
-          const acc = inward.get(id);
-          if (!acc || acc.lengthSq() < 1e-30) continue;
-          dir = acc.clone().normalize().multiplyScalar(pressureSign);
-          mag = Math.abs(g.totalForce) * a; // p · Aᵢ
-        }
-        if (mag < 1e-30) continue;
-        const prev = byNode.get(id);
-        if (!prev || mag > prev.mag) byNode.set(id, { dir, mag });
-      }
-    }
-
-    if (byNode.size === 0) return [];
-    let maxMag = 0;
-    for (const { mag } of byNode.values()) if (mag > maxMag) maxMag = mag;
-    if (maxMag < 1e-30) return [];
-
-    const up = new THREE.Vector3(0, 1, 0);
-    const arrows: {
-      pos: [number, number, number];
-      quaternion: THREE.Quaternion;
-      frac: number;
-    }[] = [];
-    for (const [id, { dir, mag }] of byNode) {
-      const e = nodeMap.get(id);
-      if (!e) continue;
-      const q = new THREE.Quaternion().setFromUnitVectors(up, dir);
-      arrows.push({
-        pos: [e.n.x, e.n.y, e.n.z],
-        quaternion: q,
-        frac: mag / maxMag,
-      });
-    }
-    return arrows;
-  }, [loadGroups, nodes, nodeMap, boundaryTriFaceIds, boundaryQuadFaceIds]);
-
-  const stepGeometry = useMemo(() => {
-    if (!stepSurface || stepSurface.triangles.length === 0) return null;
-    const { points, triangles } = stepSurface;
-    const positions = new Float32Array(triangles.length * 9);
-    const normals = new Float32Array(triangles.length * 9);
-    let pi = 0;
-    for (const [a, b, c] of triangles) {
-      const pa = points[a],
-        pb = points[b],
-        pc = points[c];
-      positions[pi] = pa[0];
-      positions[pi + 1] = pa[1];
-      positions[pi + 2] = pa[2];
-      positions[pi + 3] = pb[0];
-      positions[pi + 4] = pb[1];
-      positions[pi + 5] = pb[2];
-      positions[pi + 6] = pc[0];
-      positions[pi + 7] = pc[1];
-      positions[pi + 8] = pc[2];
-      const ax = pb[0] - pa[0],
-        ay = pb[1] - pa[1],
-        az = pb[2] - pa[2];
-      const bx = pc[0] - pa[0],
-        by = pc[1] - pa[1],
-        bz = pc[2] - pa[2];
-      let nx = ay * bz - az * by,
-        ny = az * bx - ax * bz,
-        nz = ax * by - ay * bx;
-      const len = Math.sqrt(nx * nx + ny * ny + nz * nz) || 1;
-      nx /= len;
-      ny /= len;
-      nz /= len;
-      normals[pi] = nx;
-      normals[pi + 1] = ny;
-      normals[pi + 2] = nz;
-      normals[pi + 3] = nx;
-      normals[pi + 4] = ny;
-      normals[pi + 5] = nz;
-      normals[pi + 6] = nx;
-      normals[pi + 7] = ny;
-      normals[pi + 8] = nz;
-      pi += 9;
-    }
-    return { positions, normals };
-  }, [stepSurface]);
-
-  if (nodes.length === 0 && !stepGeometry?.positions) {
+  const hasStepTessellation = !!stepSurface && stepSurface.triangles.length > 0;
+  if (nodes.length === 0 && !hasStepTessellation) {
     return null;
   }
 
@@ -957,285 +61,26 @@ export function MeshScene() {
   // only thing to show before a mesh exists). It must never paint over a solved
   // result, so it is suppressed in results mode.
   const showStepSurface =
-    !showResult &&
-    (viewRepr === "geometry" || nodes.length === 0) &&
-    !!stepGeometry;
-  // Edge overlays per representation: Geometry → none, Surface → boundary edges,
-  // Volume → every element edge, Wireframe → every element edge with no fill.
-  const showSolid = viewRepr !== "wireframe";
-  const showSurfaceEdges = viewRepr === "surface";
-  const showAllEdges = viewRepr === "volume" || viewRepr === "wireframe";
+    !showResult && (viewRepr === "geometry" || nodes.length === 0);
 
   return (
     <group>
-      {/* Undeformed solid surface — light blue-grey on light background */}
-      {!showResult && undeformedSurface && (
-        <mesh onClick={pickMode ? handleFacePick : undefined}>
-          <bufferGeometry>
-            <bufferAttribute
-              attach="attributes-position"
-              args={[undeformedSurface.positions, 3]}
-            />
-            <bufferAttribute
-              attach="attributes-normal"
-              args={[undeformedSurface.normals, 3]}
-            />
-          </bufferGeometry>
-          <meshStandardMaterial
-            color="#b8cce4"
-            side={THREE.DoubleSide}
-            wireframe={!showSolid}
-          />
-        </mesh>
+      <FemMeshLayer
+        topology={topology}
+        deformScale={deformScale}
+        showResult={showResult}
+        onFacePick={onFacePick}
+      />
+      {showResult && (
+        <ResultsColormap
+          topology={topology}
+          deformScale={deformScale}
+          onFacePick={onFacePick}
+        />
       )}
-
-      {/* Surface (boundary) edges — Surface representation */}
-      {!showResult && showSurfaceEdges && undeformedSurfaceEdgePositions && (
-        <lineSegments>
-          <bufferGeometry>
-            <bufferAttribute
-              attach="attributes-position"
-              args={[undeformedSurfaceEdgePositions, 3]}
-            />
-          </bufferGeometry>
-          <lineBasicMaterial color="#2d4a6b" />
-        </lineSegments>
-      )}
-
-      {/* All element edges — Volume and Wireframe representations */}
-      {!showResult && showAllEdges && undeformedEdgePositions && (
-        <lineSegments>
-          <bufferGeometry>
-            <bufferAttribute
-              attach="attributes-position"
-              args={[undeformedEdgePositions, 3]}
-            />
-          </bufferGeometry>
-          <lineBasicMaterial color="#2d4a6b" />
-        </lineSegments>
-      )}
-
-      {/* Deformed solid surface */}
-      {showResult && deformedSurface && (
-        <mesh onClick={pickMode ? handleFacePick : undefined}>
-          <bufferGeometry>
-            <bufferAttribute
-              attach="attributes-position"
-              args={[deformedSurface.positions, 3]}
-            />
-            <bufferAttribute
-              attach="attributes-color"
-              args={[deformedSurface.colors, 3]}
-            />
-          </bufferGeometry>
-          <meshStandardMaterial vertexColors side={THREE.DoubleSide} />
-        </mesh>
-      )}
-
-      {/* Deformed surface (boundary) edges — Surface representation */}
-      {showResult && showSurfaceEdges && deformedSurfaceEdgePositions && (
-        <lineSegments>
-          <bufferGeometry>
-            <bufferAttribute
-              attach="attributes-position"
-              args={[deformedSurfaceEdgePositions, 3]}
-            />
-          </bufferGeometry>
-          <lineBasicMaterial color="#1e3a5f" transparent opacity={0.4} />
-        </lineSegments>
-      )}
-
-      {/* Deformed element edges — Volume and Wireframe representations */}
-      {showResult && showAllEdges && deformedEdgePositions && (
-        <lineSegments>
-          <bufferGeometry>
-            <bufferAttribute
-              attach="attributes-position"
-              args={[deformedEdgePositions, 3]}
-            />
-          </bufferGeometry>
-          <lineBasicMaterial color="#1e3a5f" transparent opacity={0.4} />
-        </lineSegments>
-      )}
-
-      {/* Undeformed geometry overlay — shows original surface edges as reference over deformed result */}
-      {showResult &&
-        showUndeformedOverlay &&
-        undeformedSurfaceEdgePositions && (
-          <lineSegments>
-            <bufferGeometry>
-              <bufferAttribute
-                attach="attributes-position"
-                args={[undeformedSurfaceEdgePositions, 3]}
-              />
-            </bufferGeometry>
-            <lineBasicMaterial color="#6b8cad" transparent opacity={0.5} />
-          </lineSegments>
-        )}
-
-      {/* BC face highlights — persistent coloured overlay for all committed BC faces */}
-      {!showResult &&
-        bcFaceHighlights?.map((h, i) => (
-          <mesh key={`bc-face-${h.groupId}-${i}`} renderOrder={1}>
-            <bufferGeometry>
-              <bufferAttribute
-                attach="attributes-position"
-                args={[h.positions, 3]}
-              />
-            </bufferGeometry>
-            <meshBasicMaterial
-              color="#dc2626"
-              transparent
-              opacity={pickTargetGroupId === h.groupId ? 0.45 : 0.25}
-              depthTest={false}
-              side={THREE.DoubleSide}
-            />
-          </mesh>
-        ))}
-
-      {/* Load face highlights — persistent coloured overlay for all committed load faces */}
-      {!showResult &&
-        loadFaceHighlights?.map((h, i) => (
-          <mesh key={`load-face-${h.groupId}-${i}`} renderOrder={1}>
-            <bufferGeometry>
-              <bufferAttribute
-                attach="attributes-position"
-                args={[h.positions, 3]}
-              />
-            </bufferGeometry>
-            <meshBasicMaterial
-              color="#d97706"
-              transparent
-              opacity={pickTargetGroupId === h.groupId ? 0.45 : 0.25}
-              depthTest={false}
-              side={THREE.DoubleSide}
-            />
-          </mesh>
-        ))}
-
-      {/* Pending faces — accumulated via shift-click, same colour as selection but slightly dimmer */}
-      {pendingFacePositions && (
-        <mesh renderOrder={2}>
-          <bufferGeometry>
-            <bufferAttribute
-              attach="attributes-position"
-              args={[pendingFacePositions, 3]}
-            />
-          </bufferGeometry>
-          <meshBasicMaterial
-            color="#e05533"
-            transparent
-            opacity={0.45}
-            depthTest={false}
-            side={THREE.DoubleSide}
-          />
-        </mesh>
-      )}
-
-      {/* Selected face highlight — the latest picked face (brightest) */}
-      {selectedFacePositions && (
-        <mesh renderOrder={3}>
-          <bufferGeometry>
-            <bufferAttribute
-              attach="attributes-position"
-              args={[selectedFacePositions, 3]}
-            />
-          </bufferGeometry>
-          <meshBasicMaterial
-            color="#e05533"
-            transparent
-            opacity={0.65}
-            depthTest={false}
-            side={THREE.DoubleSide}
-          />
-        </mesh>
-      )}
-
-      {/* BC markers — triangular fixed-support symbols (3-sided cone, apex at face, base outward) */}
-      {!showResult && bcMarkerData && (
-        <group position={bcMarkerData.pos} quaternion={bcMarkerData.quaternion}>
-          <mesh position={[0, -modelSize * 0.075, 0]}>
-            <coneGeometry args={[modelSize * 0.09, modelSize * 0.15, 3]} />
-            <meshStandardMaterial color="#dc2626" />
-          </mesh>
-          {/* Backing strip representing the fixed wall */}
-          <mesh
-            position={[0, -modelSize * 0.165, 0]}
-            rotation={[Math.PI / 2, 0, 0]}
-          >
-            <planeGeometry args={[modelSize * 0.22, modelSize * 0.03]} />
-            <meshStandardMaterial color="#dc2626" side={THREE.DoubleSide} />
-          </mesh>
-        </group>
-      )}
-
-      {/* Resultant load arrows — one per force/pressure group, cylinder shaft + cone head */}
-      {!showResult &&
-        loadDisplay === "resultant" &&
-        loadArrows.map((arrow, i) => {
-          const shaftLen = modelSize * 0.22;
-          const headLen = modelSize * 0.09;
-          const shaftR = modelSize * 0.012;
-          const headR = modelSize * 0.038;
-          return (
-            <group key={i} position={arrow.pos} quaternion={arrow.quaternion}>
-              <mesh position={[0, shaftLen / 2, 0]}>
-                <cylinderGeometry args={[shaftR, shaftR, shaftLen, 8]} />
-                <meshStandardMaterial color="#d97706" />
-              </mesh>
-              <mesh position={[0, shaftLen + headLen / 2, 0]}>
-                <coneGeometry args={[headR, headLen, 8]} />
-                <meshStandardMaterial color="#d97706" />
-              </mesh>
-            </group>
-          );
-        })}
-
-      {/* Per-node load arrows — one per loaded node, length scaled by the load
-          that node carries (relative to the largest in the model) */}
-      {!showResult &&
-        loadDisplay === "nodal" &&
-        nodalLoadArrows.map((arrow, i) => {
-          // Shortest arrows stay at 35% of the full length so light-loaded
-          // nodes remain visible; radii are fixed so arrows read as a field.
-          const len = modelSize * 0.13 * (0.35 + 0.65 * arrow.frac);
-          const shaftLen = len * 0.68;
-          const headLen = len * 0.32;
-          const shaftR = modelSize * 0.006;
-          const headR = modelSize * 0.018;
-          return (
-            <group key={i} position={arrow.pos} quaternion={arrow.quaternion}>
-              <mesh position={[0, shaftLen / 2, 0]}>
-                <cylinderGeometry args={[shaftR, shaftR, shaftLen, 6]} />
-                <meshStandardMaterial color="#d97706" />
-              </mesh>
-              <mesh position={[0, shaftLen + headLen / 2, 0]}>
-                <coneGeometry args={[headR, headLen, 8]} />
-                <meshStandardMaterial color="#d97706" />
-              </mesh>
-            </group>
-          );
-        })}
-
-      {/* STEP surface mesh — visible in geometry mode only; replaced by FEM mesh in mesh mode */}
-      {showStepSurface && stepGeometry && (
-        <mesh>
-          <bufferGeometry>
-            <bufferAttribute
-              attach="attributes-position"
-              args={[stepGeometry.positions, 3]}
-            />
-            <bufferAttribute
-              attach="attributes-normal"
-              args={[stepGeometry.normals, 3]}
-            />
-          </bufferGeometry>
-          <meshStandardMaterial
-            color="#7a9bbf"
-            side={THREE.DoubleSide}
-            wireframe={!showSolid}
-          />
-        </mesh>
+      <BoundaryConditionLayer topology={topology} showResult={showResult} />
+      {showStepSurface && (
+        <GeometryLayer wireframe={viewRepr === "wireframe"} />
       )}
     </group>
   );
